Create old_pca
This is the original PCA code we used to create the visualization for the presentation. It does not include any calls to the database directory as I am completely unable to do it from home, and unable to log in to database in the lab (sorting it out with Support atm). 

Will commit the newer file which calls from the database once I have tested it.